# Conflicts:
#	iosgl/iosgl.xcworkspace/xcuserdata/zhao.xcuserdatad/UserInterfaceState.xcuserstate

+++ iosgl/iosgl/bt/utils/TimeUtil.h
@@ -20,8 +20,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface TimeUtil : NSObject
 + (instancetype)default;
-@property (nonatomic, assign) float    time  ;
- -(int)getTimer;
+@property (nonatomic, assign) double    time  ;
+ 
+-(double)getTimer;
 @end
 
 NS_ASSUME_NONNULL_END
